requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
wikipedia==1.4.0
pydantic==2.5.0
typing-extensions==4.9.0
annotated-types==0.6.0
google-generativeai==0.3.2